fix: user message didn't start a session if a campus message preceded it
the last message getter was getting the time for any last message, and when a recent campus last message was found (which didn't have any session) its time was compared with curTime, and session was perceived to be valid.
However, there was not session to be checked in the first place if the previous message didn't belong to any session.

This is why, the last message used to compare time difference with now, needs to belong to a valid session