fix(scholarship):  update scholarship schema to allow saving documents without eligibility criteria
eligibility criteria was added as a required field in the scholarship schema in a recent change. However, since this was not an enforced field in the schema previously, many documents had no `eligibility_criteria` field. This caused a db error when any update/create operation was triggered on the collection